+++ gulp/tasks/copy.js
@@ -19,6 +19,13 @@ gulp.task( 'copy:icons', function() {
     .pipe( gulp.dest( config.icons.dest ) );
 } );
 
+gulp.task( 'copy:libjs', function() {
+  return gulp.src( config.libjs.src )
+    .pipe( $.changed( config.libjs.dest ) )
+    .on( 'error', handleErrors )
+    .pipe( gulp.dest( config.libjs.dest ) );
+} );
+
 gulp.task( 'copy:vendorjs', function() {
   return gulp.src( config.vendorjs.src )
     .pipe( $.changed( config.vendorjs.dest ) )
@@ -30,6 +37,7 @@ gulp.task( 'copy',
   [
     'copy:files',
     'copy:icons',
+    'copy:libjs',
     'copy:vendorjs'
   ]
 );